+++ src/monolithic/index.js
@@ -1,7 +1,6 @@
 
 import { StyleSheet } from 'glamor/lib/sheet'
 import hash from '../hash'
-import { addPx, hyphenate } from '../util'
 
 const options = {
 }
@@ -93,6 +92,47 @@ const createRule = (selector, decs, media) => {
   const css = media ? `${media}{${rule}}` : rule
   return css
 }
+
+export const hyphenate = (str) => ('' + str)
+  .replace(/[A-Z]|^ms/g, '-$&')
+  .toLowerCase()
+
+export const addPx = (prop, value) => {
+  if (typeof value !== 'number') return value
+  if (unitlessProps.indexOf(prop) > -1) return value
+  return value + 'px'
+}
+
+const unitlessProps = [
+  'animationIterationCount',
+  'boxFlex',
+  'boxFlexGroup',
+  'boxOrdinalGroup',
+  'columnCount',
+  'flex',
+  'flexGrow',
+  'flexPositive',
+  'flexShrink',
+  'flexNegative',
+  'flexOrder',
+  'gridRow',
+  'gridColumn',
+  'fontWeight',
+  'lineClamp',
+  'lineHeight',
+  'opacity',
+  'order',
+  'orphans',
+  'tabSize',
+  'widows',
+  'zIndex',
+  'zoom',
+  'fillOpacity',
+  'stopOpacity',
+  'strokeDashoffset',
+  'strokeOpacity',
+  'strokeWidth'
+]
 
 cxs.reset = reset
 cxs.getCss = getCss
